feat(log-table): add expandable quota details and refactor price calculation (#728)
Introduced an expandable row to display detailed quota information, including original and actual pricing, group ratio, and calculation breakdown. Refactored price calculation logic for clarity and reusability. Updated `TableRow` component to integrate new features.
